mv parser from utils to main
